examples/fastapi-app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
prometheus-client==0.21.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-logging==0.48b0
opentelemetry-exporter-otlp-proto-http==1.27.0
sentry-sdk[fastapi]==2.19.0
